refactor(prompt): extract reference merging to prompt-reference-utils to resolve file and function length lint errors
