+++ src/components/organisms/Workbench.tsx
@@ -7,7 +7,8 @@ import { useEvolution } from "../../hooks/useEvolution"
 import { usePromptInjector } from "../../hooks/usePromptInjector"
 import { useWorkbench } from "../../hooks/useWorkbench"
 import type { PromptSegment } from "../../lib/db-schema"
-import { mergePromptSegments, mergeReferences } from "../../lib/prompt-utils"
+import { mergePromptSegments } from "../../lib/prompt-utils"
+import { mergeReferences } from "../../lib/prompt-reference-utils"
 import { type AlertType } from "../molecules/ConnectionAlert"
 import {
   evolveTargetCard,
